processview的样式调整

+++ DDCOfflineStore/Classes/Models/Local/ContractModel/DDCContractModel.swift
@@ -25,7 +25,7 @@ class DDCContractModel: Mappable {
     
     var contractType: DDCContractType?
     var courseType: DDCCourseType = .regular
-    var contractPrice: Int?
+    var contractPrice: Double?
     var createdUsername: String?
     var signedUsername: String?
     var responsibleUsername: String?
